+++ packages/composites/reference-transform/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @chainlink/reference-transform-adapter
 
+## 1.2.60
+
+### Patch Changes
+
+- @chainlink/ea@1.3.59
+
 ## 1.2.59
 
 ### Patch Changes